# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestão Jurídica",
        "description": "Sistema de Gestão Jurídica",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/processos": {
            "get": {
                "tags": [
                    "Processos"
                ],
                "summary": "Listar Processos",
                "operationId": "processo/index",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "the page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "order  accepts 'asc' or 'desc'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/processo": {
            "post": {
                "tags": [
                    "Processos"
                ],
                "summary": "Cadastrar Processo",
                "operationId": "processo/store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codigo"
                                ],
                                "properties": {
                                    "codigo": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Matéria Teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/processo/{id}": {
            "get": {
                "tags": [
                    "Processos"
                ],
                "summary": "Pesquisar Processo",
                "operationId": "processo/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Processo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Processos"
                ],
                "summary": "Editar Processo",
                "operationId": "processo/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Processo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "codigo"
                                ],
                                "properties": {
                                    "codigo": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Matéria Teste"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Processos"
                ],
                "summary": "Excluir Processo",
                "operationId": "processo/destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Processo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}